pyproject.toml
[build-system]
requires = ["hatchling>=1.29.0"]
build-backend = "hatchling.build"

[project]
name = "llm-observe-proxy"
version = "0.1.0"
description = "OpenAI-compatible LLM proxy with SQLite request capture, observability, and an admin UI."
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "LLM Observe Proxy Contributors" }]
keywords = ["llm", "openai", "proxy", "observability", "sqlite", "fastapi", "admin-ui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Debuggers",
]
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown-it-py>=4.0.0",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[project.urls]
Homepage = "https://github.com/shamitv/llm-observe-proxy"
Repository = "https://github.com/shamitv/llm-observe-proxy"
Issues = "https://github.com/shamitv/llm-observe-proxy/issues"
Changelog = "https://github.com/shamitv/llm-observe-proxy/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "build>=1.5.0",
    "playwright>=1.59.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "respx>=0.23.1",
    "ruff>=0.15.12",
    "twine>=6.2.0",
]

[project.scripts]
llm-observe-proxy = "llm_observe_proxy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_observe_proxy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
